% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_color_groups.R
\name{gt_color_groups}
\alias{gt_color_groups}
\title{Color \code{gtsummary} variable group headers}
\usage{
gt_color_groups(tbl, color = "#B31B1B", fill = "grey95")
}
\arguments{
\item{tbl}{A \code{gtsummary} table.}

\item{color}{Text color for variable group rows.}

\item{fill}{Fill color for variable group rows.}
}
\value{
A styled \code{gt} table.
}
\description{
Color \code{gtsummary} variable group headers
}
