% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexcolor.R
\name{hexcolor}
\alias{hexcolor}
\title{Hex color picker app}
\usage{
hexcolor(value = "#FFFFFF", height = "200px", viewer = shiny::paneViewer())
}
\arguments{
\item{value}{Initial color value.}

\item{height}{Plot preview height.}

\item{viewer}{Viewer to launch the app in. Defaults to the RStudio Viewer pane.}
}
\value{
A Shiny app is launched.
}
\description{
Opens a small Shiny app with a color picker.
}
