% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_data}
\alias{save_data}
\title{Save a dataframe as an .rds file}
\usage{
save_data(data, file, loc = "data")
}
\arguments{
\item{data}{Dataframe.}

\item{file}{File name for saved object (combined with current date).}

\item{loc}{Destination in R project. Defaults to \code{data} folder.}
}
\description{
Save a dataframe as an .rds file
}
