% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_tbl}
\alias{save_tbl}
\title{Save a table as an .rds file}
\usage{
save_tbl(data, file, loc = "reports/tbls")
}
\arguments{
\item{data}{Dataframe.}

\item{file}{File name for saved object (combined with current date).}

\item{loc}{Destination in R project. Defaults to \code{reports/tbls} folder.}
}
\description{
Save a table as an .rds file
}
