% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_button.R
\name{tbl_button}
\alias{tbl_button}
\title{Add download buttons for a gtsummary table in an HTML file}
\usage{
tbl_button(
  tbl,
  name,
  dir = here::here("reports", "downloads"),
  label = "Download Table",
  color_border = "#6c757d",
  color_bg = "#f8f9fa",
  color_text = "inherit",
  color_option_border = "#ced4da",
  color_option_bg = "white",
  color_option_text = "inherit",
  color_option_hover_bg = "#e9ecef",
  color_option_hover_text = "inherit"
)
}
\arguments{
\item{tbl}{A gtsummary table.}

\item{name}{A single non-empty string used as the file name.}

\item{dir}{Directory where exported files should be saved.}

\item{label}{Button label.}

\item{color_border}{Button border color.}

\item{color_bg}{Button background color.}

\item{color_text}{Button text color.}

\item{color_option_border}{Option border color.}

\item{color_option_bg}{Option background color.}

\item{color_option_text}{Option text color.}

\item{color_option_hover_bg}{Option hover background color.}

\item{color_option_hover_text}{Option hover text color.}
}
\value{
An HTML tag list containing the download button. Pastes raw HTML into console if used outside of markdown render.
}
\description{
Add download buttons for a gtsummary table in an HTML file
}
