% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_code}
\alias{toggl_code}
\title{Start a Toggl timer for coding time}
\usage{
toggl_code(desc = "Code/Reports", tags = basename(dirname(here::here())))
}
\arguments{
\item{desc}{Timer description. Defaults to the active project name.}

\item{tags}{Toggl tag. Defaults to the parent folder name.}
}
\description{
Start a Toggl timer for coding time
}
