% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggl.R
\name{toggl_research}
\alias{toggl_research}
\title{Start a Toggl timer for research time}
\usage{
toggl_research(
  desc = "Research/Writing",
  tags = basename(dirname(here::here()))
)
}
\arguments{
\item{desc}{Timer description. Defaults to the active project name.}

\item{tags}{Toggl tag. Defaults to the parent folder name.}
}
\description{
Start a Toggl timer for research time
}
