% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_style_groups.R
\name{gtsummary_style_groups}
\alias{gtsummary_style_groups}
\title{Style \code{gtsummary} variable group headers}
\usage{
gtsummary_style_groups(tbl, indent = 8L)
}
\arguments{
\item{tbl}{A \code{gtsummary} table.}

\item{indent}{Indentation size for level rows.}
}
\value{
A styled \code{gtsummary} table.
}
\description{
Style \code{gtsummary} variable group headers
}
